# openapi.yaml
openapi: 3.1.0
info:
  title: Halden Threat Detection
  version: "1.0.0"
  description: |
    Serves malware detections raised by Halden's backup scanning pipeline.

    Calls carry a short-lived signed bearer token minted by the platform. The
    service verifies the token and reads detections for the scope the token
    carries: a token with `tenant_id` reads that tenant; a token with the
    `platform:aggregate` scope reads across the estate, which is the scope the
    nightly rollup runs with. The service runs in the `halden` namespace on
    port 8000.
servers:
  - url: http://halden-threat-detection.halden.svc.cluster.local:8000
components:
  securitySchemes:
    internalToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 token signed with the shared internal secret. Verified for
        signature, `exp`, `iss` (halden-identity) and `aud`
        (halden-threat-detection).
  parameters:
    RequestId:
      name: X-Halden-Request-ID
      in: header
      required: false
      description: Telemetry only; carries no authority.
      schema:
        type: string
    ClientVersion:
      name: X-Halden-Client-Version
      in: header
      required: false
      description: Telemetry only; carries no authority.
      schema:
        type: string
    Locale:
      name: X-Halden-Locale
      in: header
      required: false
      description: Telemetry only; carries no authority.
      schema:
        type: string
  schemas:
    Detection:
      type: object
      required: [id, resource_name, malware_family, severity, detected_at, status]
      properties:
        id: { type: string }
        resource_name: { type: string }
        malware_family: { type: string }
        severity:
          type: string
          enum: [low, medium, high, critical]
        detected_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [quarantined, under_review, remediated]
    ScanList:
      type: object
      required: [scope, detections, summary]
      properties:
        scope:
          type: string
          description: The tenant this response is scoped to, or "estate" for an estate-wide caller.
        detections:
          type: array
          items:
            $ref: '#/components/schemas/Detection'
        summary:
          type: object
          additionalProperties:
            type: integer
    ScanSummary:
      type: object
      required: [scope, total, by_severity, top_resources]
      properties:
        scope:
          type: string
          description: The tenant this summary is scoped to, or "estate" for an estate-wide caller.
        total:
          type: integer
        by_severity:
          type: object
          additionalProperties:
            type: integer
        top_resources:
          type: array
          items:
            type: object
            properties:
              tenant_id: { type: string }
              resource_name: { type: string }
              malware_family: { type: string }
security:
  - internalToken: []
paths:
  /v1/scans:
    get:
      summary: List detections for the caller's scope
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          description: Detections for the caller's scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanList'
        '401':
          description: Missing or invalid token
        '403':
          description: Token carries no read scope
  /v1/scans/summary:
    get:
      summary: Detection totals and recent items for the caller's scope
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          description: Summary for the caller's scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanSummary'
        '401':
          description: Missing or invalid token
        '403':
          description: Token carries no read scope
  /v1/scans/{scan_id}:
    get:
      summary: Fetch one detection within the caller's scope
      parameters:
        - name: scan_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          description: The detection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
        '401':
          description: Missing or invalid token
        '403':
          description: Token carries no read scope
        '404':
          description: No such detection within the caller's scope
  /healthz:
    get:
      summary: Liveness and readiness probe
      responses:
        '200':
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string